including nudging scripts
