+++ pyproject.toml
@@ -29,3 +29,5 @@ make_ic_template    = "pyroms_tools.initial.ic_file_template:main"
 make_ic_build       = "pyroms_tools.initial.ic_file_build:main"
 make_boundary_template    = "pyroms_tools.boundary.boundary_file_template:main"
 make_boundary_build       = "pyroms_tools.boundary.boundary_file_build:main"
+make_nudging_template    = "pyroms_tools.nudging.nudging_file:main"
+make_nudging_build       = "pyroms_tools.nudging.make_climatology:main"
