+++ pyroms_tools/utils/utils.py
@@ -2,13 +2,12 @@ import ast
 import xarray as xr
 import numpy as np
 from scipy import interpolate as itp
+from interpolate import interpolate as itp1
 import netCDF4 as nc
-try:
-    import pyroms
-except:
-    pass
-# from metpy import interpolate as metitp
-
+import datetime  as dtt
+import gsw
+import pyroms
+import xesmf as xe
 
 def _get_dict_paths(fpath):
     """Read a text file with a dictionary and returns
@@ -47,6 +46,12 @@ def get_ref_dicts(fpath, ref, src_path, out_path):
             temp[bla] = dicts[bla]
     print(temp)
     return temp
+
+
+def interpolate_horizontal(source_grid: xr.Dataset, target_grid: xr.Dataset) -> xr.Dataset:
+    """Interpolate data from source grid to target grid using bilinear method."""
+    regridder = xe.Regridder(source_grid, target_grid, 'bilinear', extrap_method='nearest_s2d')
+    return regridder(source_grid)
 
 # def _nn_interpolation(xc,yc,zc,xo,yo):
 #     """Wrapper for metpy natural neighbor interpolation method.
@@ -374,3 +379,288 @@ class nudgcoef():
         # close
         fid.close()
         return None
+
+
+
+
+class interpolationGlorys2Roms(object):
+    def __init__(self, fgrid, fpath, tslice=None, load=False):
+        self.dsgrid   = xr.open_dataset(fgrid)
+        self.dssource = xr.open_mfdataset(fpath)
+        self.dssource  = self.dssource.drop_duplicates(dim='time')
+        if tslice is not None:
+            tstart=tslice[0]
+            tfinal =tslice[1]
+            tstart_num = dtt.datetime.strptime(tstart, '%Y-%m-%dT%H:%M:%S')
+            tfinal_num = dtt.datetime.strptime(tfinal, '%Y-%m-%dT%H:%M:%S')
+
+            self.dssource = self.dssource.sel(time=slice(tslice[0], tslice[1]))
+
+        if load:
+            self.dssource.load()
+
+    def monthly_mean(self):
+        return self.dssource.groupby('time.month').mean(dim='time')
+        
+
+    def set_coords(self, itime, gtype='rho', ztype='rho', time_type=None):
+        """
+        Set coordinates of glorys to conform with C grid from roms and select
+        a given time step of the dataset.  The two options to create the fields
+        are monthly means (not climatological) or a given date -- see itime
+        argument for more information
+        
+
+        itime (integer or datetime):selects the month /date of the interpolation.
+        time_type (string):
+            'monthly': a monthly mean is calculated and the integer
+                  index should be used. For example, if march and may are included in
+                  the source netcdf, march and may will be index 0 and 1, respectively.
+            None     : uses the date in itime
+        gtype (string): ['rho', 'u', 'v']:
+            remap horizontal lon, lat to rho, u or v points in Arakawa C-grid
+        ztype: ['rho','w']:
+            remap vertical positions to rho, or w points in Arakawa C-grid
+        """
+        dsgrid = self.dsgrid.copy()
+
+        # renaming for interpolation standards ds1.roms.interpolate
+        dsgrid = dsgrid.rename({f'lon_{gtype}':'lon', f'lat_{gtype}':'lat'})
+        dsgrid = dsgrid.set_coords(['lon', 'lat'])
+
+        # select and set up time
+        if time_type is None:
+            self.dssource1 = self.dssource
+            ds0 = self.dssource1.sel(time=itime, method='nearest')
+
+        elif time_type == 'monthly':
+            self.dssource1 = self.monthly_mean()
+            self.dssource1 = self.dssource1.rename(month='time')
+            ds0 = self.dssource1.isel(time=itime)
+        else:
+            raise IOError("time_type shoud be None or 'monthly'")
+        
+
+        self.ds1 = interpolate_horizontal(ds0, dsgrid)
+        a = self.ds1
+
+        # glorys interpolated horizontally to roms
+        self.zz   = -np.tile(a.depth.values[:,None,None], (1, *a.lon.shape) )
+        self.lonz = np.broadcast_to(a.lon.values, (ds0.depth.size, *a.lon.shape))
+        self.latz = np.broadcast_to(a.lat.values, (ds0.depth.size, *a.lat.shape))
+
+
+        self.zrho   = s2z(self.dsgrid, ztype).transpose(f's_{ztype}',f'eta_{ztype}', f'xi_{ztype}')
+        
+        self.zu     =  self.zrho#.rename(lon_rho='lon_u', lat_rho='lat_u', eta_rho='eta_u', xi_rho='xi_u')
+        self.zu     = self.zu[:,:,:-1]
+        # self.zu.coords['lon_u'].values = self.dsgrid.lon_u
+        # self.zu.coords['lat_u'].values = self.dsgrid.lat_u
+
+        self.zv     =  self.zrho#.rename(lon_rho='lon_v', lat_rho='lat_v', eta_rho='eta_v', xi_rho='xi_v')
+        self.zv     = self.zv[:,:-1,:]
+        # self.zv.coords['lon_v'].values = self.dsgrid.lon_v
+        # self.zv.coords['lat_v'].values = self.dsgrid.lat_v
+
+        
+        if gtype=='u':
+            self.zg = self.zu
+            self.long = np.broadcast_to(dsgrid['lon'].values, self.zu.shape)
+            self.latg = np.broadcast_to(dsgrid['lat'].values, self.zu.shape)
+
+        elif gtype=='v':
+            self.zg = self.zv
+            self.long = np.broadcast_to(dsgrid['lon'].values, self.zv.shape)
+            self.latg = np.broadcast_to(dsgrid['lat'].values, self.zv.shape)
+
+        elif gtype=='rho':
+            self.zg = self.zrho
+            self.long = np.broadcast_to(dsgrid['lon'].values, self.zrho.shape)
+            self.latg = np.broadcast_to(dsgrid['lat'].values, self.zrho.shape)
+
+        else:
+            IOError('gtype must be u,v or rho')
+
+
+        return self.dssource1['time']
+
+
+    def calculate_geostrophy(self):
+        ds = self.dssource
+        xm, ym = np.meshgrid(ds.latitude.values,
+                            ds.longitude.values,indexing='ij')
+
+        f = gsw.f(ym)
+        g = 9.8
+
+
+
+        ds['f'] = (['latitude', 'longitude'], f)
+        ds['psi'] = ds['zos'] * 9.8/ ds['f']
+
+        ds['dpsi_x'] = (['time', 'latitude', 'longitude'], np.gradient(ds['psi'].values, axis=2))
+        ds['dpsi_y'] = (['time', 'latitude', 'longitude'], np.gradient(ds['psi'].values, axis=1))
+        ds.drop('psi')
+
+
+        # calcualte distances
+        lons = ds.longitude.values
+        lats = ds.latitude.values
+
+        # provisory dy dx
+        dy = gsw.distance(xm,ym,axis=1)
+        dx = gsw.distance(xm,ym,axis=0)
+
+        # cumulative sum 
+        x = np.cumsum(dx, axis=1)
+        y = np.cumsum(dy, axis=0)
+
+        # recalculating distances to preserve dimensions
+        x = np.insert(x, 0, values=0, axis=0)
+        y = np.insert(y, 0, values=0, axis=1)
+
+        dx = np.gradient(x, axis=1)
+        dy = np.gradient(y, axis=0)
+
+        # set in xarray dataset 
+        ds['dx'] = (['latitude', 'longitude'], dx)
+        ds['dy'] = (['latitude', 'longitude'], dy)
+
+        # calculate geostrophic velocities
+        ds['ug'] = -ds['dpsi_y']/ds['dy']
+        ds['vg'] = ds['dpsi_x']/ds['dx']
+
+
+        # plt.close('all')
+        # fig = plt.figure(figsize=[20,20])
+        # plt.contourf(xm,ym,ds['zos'][0], levels=30,zorder=0, cmap=plt.cm.jet)
+        # plt.quiver(xm[::2,::2],ym[::2,::2], ds['ug'][0,::2,::2], ds['vg'][0,::2,::2], scale=20, width=0.0005)
+        # plt.savefig('bla.png', dpi=300)
+
+    def interpolate3d(self,varb):
+        zs = self.zg
+        outfield = np.zeros(self.zg.shape)
+
+        xm = self.lonz
+        ym = self.latz
+        zm = self.zz[::-1]
+        varb = self.ds1[varb]
+        varb = varb.bfill('xi_rho')
+        varb = varb.ffill('xi_rho')
+        varb = varb.bfill('eta_rho')
+        varb = varb.ffill('eta_rho')
+
+        varb = varb.bfill('depth')
+
+        varb = varb.values[::-1]
+
+        nzs = zs.shape[0]
+        nz, ny, nx = xm.shape
+        aux = itp1.interp3d_along_axis0(zs,
+                                    xm,ym,zm,
+                                    varb,
+                                    nzs,nx,ny,nz)
+        aux = np.where(aux==9999.0, np.nan, aux)
+        outfield[...] = aux
+        return outfield
+
+def nearest_interpolation(dsgrid, varb, hgrid='rho', barotropic=False):
+    if not barotropic:
+        dsgrid[varb] = dsgrid[varb].bfill('s_rho')
+        dsgrid[varb] = dsgrid[varb].ffill('s_rho')
+
+    dsgrid[varb] = dsgrid[varb].ffill(f'xi_{hgrid}')
+    dsgrid[varb] = dsgrid[varb].bfill(f'xi_{hgrid}')
+    dsgrid[varb] = dsgrid[varb].ffill(f'eta_{hgrid}')
+    dsgrid[varb] = dsgrid[varb].bfill(f'eta_{hgrid}')
+    return dsgrid
+
+
+
+    def s2z(self, coords, transform=ut.vtransform2,
+                        zeta=None,
+                        trans_coords=None):
+        """coords is either 'rho', 'u', or 'v'"""
+        z = compute_depth_layers(self._obj, coords, transform=transform,
+                             zeta=zeta, trans_coords=trans_coords)
+        return z
+
+    
+
+def vtransform2(zeta, sigma, C, hc, h):
+    S = (hc * sigma + h*C) / (hc +h)
+    z = zeta + (zeta + h) * S
+    return z       
+
+
+def s2z(ds, coords, transform=vtransform2,
+                    zeta=None,
+                    trans_coords=None):
+    """coords is either 'rho', 'u', or 'v'"""
+    z = compute_depth_layers(ds, coords, transform=transform,
+                            zeta=zeta, trans_coords=trans_coords)
+    return z
+
+    
+
+def compute_depth_layers(ds, coords, transform=vtransform2,
+                        zeta=None,
+                        trans_coords=None):
+    """
+    Compute depth layers for a ROMS model.
+
+    Parameters:
+    - ds (xarray.Dataset): The input dataset containing ROMS model variables.
+    - coords (str): The type of vertical coordinate to use for computing the depth layers.
+                    Possible values: 'w' for W-coordinate, 'rho' for Rho-coordinate.
+    - transform (function, optional): The vertical coordinate transformation function.
+                                      Default: vtransform2 (ROMS vtransform2 function).
+    - zeta (xarray.DataArray or None, optional): The free surface elevation.
+                                                 Default: None (assume zeta = 0).
+    - trans_coords (tuple or None, optional): The desired order of dimensions for the output depth layers.
+                                              Default: None (no transpose).
+
+    Returns:
+    - z (xarray.DataArray): The computed depth layers as an xarray DataArray.
+
+    Notes:
+    - This function assumes that the input dataset (`ds`) contains the necessary variables for the depth calculation.
+    - The depth calculation is performed using the specified vertical coordinate transformation function (`transform`).
+    - If `zeta` is provided, it represents the free surface elevation; otherwise, it is assumed to be zero.
+    - The computed depth layers are returned as an xarray DataArray.
+    - If `trans_coords` is specified, the output depth layers will be transposed accordingly.
+
+    more information in: https://www.myroms.org/wiki/Vertical_S-coordinate
+    """    
+    
+
+    assert coords in ['w','rho']
+    assert transform.__name__ in ['vtransform2']
+
+    if zeta is not None:
+        zeta = zeta
+    else:
+        zeta = 0
+
+    if (transform.__name__ == 'vtransform2') & (coords=='rho'):
+        hc    = ds.hc
+        sigma = ds.s_rho
+        C     = ds.Cs_r
+        h     = ds.h
+
+        z = transform(zeta, sigma, C, hc, h)
+    elif (transform.__name__ == 'vtransform2') & (coords=='w'):
+        hc    = ds.hc
+        sigma = ds.s_w
+        C     = ds.Cs_w
+        h     = ds.h
+
+        z = transform(zeta, sigma, C, hc, h)
+
+    if trans_coords:
+        z = z.transpose(*trans_coords)
+
+    return z
+
+
+
